requirements.txt
Flask
Flask-Caching
Flask-CORS
flask-limiter
python-dotenv
pytest
pytest-flask
pytest-cov
flask-swagger-ui
requests
pandas
scikit-learn
numpy
astral
gunicorn
pytz
schedule
Pillow
Werkzeug
opencv-python
scipy